Refactored CancelAbilities method and node
